Create README.md
#CSVParser
CSVParser is a stream based CSV library written in Swift. It uses InputStream to parse a CSV file and OutputStream to write CSV data to a file. This way it doesn't keep everything in memory while working with big CSV files. It also supports BOM for UTF-8, UTF-16 and UTF-32 text encodings.

CSVParser is battle tested in the CSV import and export of Finances.

Features
Stream based CSV parser and writer
Complete documentation
Unit tested